Create chapter1-ex2
Chapter 1-Exercise 2:
 Having listed the alternatives, next we can calculate the carbon emissions for each of them. Below you'll find the distances between the ports in kilometers in a five-by-five table.

PAN	AMS	CAS	NY	HEL
PAN	 0	8943	8019	3652	10545
AMS	8943	0	2619	6317	2078
CAS	8019	2619	0	5836	4939
NY	3652	6317	5836	0	7825
HEL	10545	2078	4939	7825	0
Let's assume that the boat is relatively modern and produces 0.020 kg of CO2 emissions per kilometer for the amount of pineapples that we are shipping. Thus, you can calculate the emissions caused by traveling from Panama to Amsterdam by first looking up the distance in the first row, second column of the table (highlighted in the above table): 8943 km, and then multiplying this with 0.020kg/km to get 178.9 kg.

Building on the previous solution, modify the code so that it finds the route with minimum carbon emissions and prints it out. Again, the program should work for any number of ports. You can assume that the distances between the ports are given in an array of the appropriate size so that the distance between ports i and j is found in D[i][j].